+++ android/src/main/java/com/reactnativestripesdk/CardFieldView.kt
@@ -6,6 +6,7 @@ import android.graphics.Typeface
 import android.os.Build
 import android.text.Editable
 import android.text.TextWatcher
+import android.util.Log
 import android.widget.FrameLayout
 import com.facebook.react.bridge.ReadableMap
 import com.facebook.react.uimanager.ThemedReactContext
@@ -21,6 +22,7 @@ import com.stripe.android.view.CardInputListener
 import com.stripe.android.view.CardInputWidget
 import com.stripe.android.view.CardValidCallback
 import com.stripe.android.view.StripeEditText
+import java.lang.Exception
 
 class CardFieldView(context: ThemedReactContext) : FrameLayout(context) {
   private var mCardWidget: CardInputWidget = CardInputWidget(context)
@@ -90,7 +92,8 @@ class CardFieldView(context: ThemedReactContext) : FrameLayout(context) {
       cardInputWidgetBinding.cardNumberEditText,
       cardInputWidgetBinding.cvcEditText,
       cardInputWidgetBinding.expiryDateEditText,
-      cardInputWidgetBinding.postalCodeEditText)
+      cardInputWidgetBinding.postalCodeEditText
+    )
 
     textColor?.let {
       for (editTextBinding in bindings) {
@@ -106,6 +109,7 @@ class CardFieldView(context: ThemedReactContext) : FrameLayout(context) {
       for (editTextBinding in bindings) {
         editTextBinding.setHintTextColor(Color.parseColor(it))
       }
+      setCardBrandTint(Color.parseColor(it))
     }
     fontSize?.let {
       for (editTextBinding in bindings) {
@@ -149,6 +153,19 @@ class CardFieldView(context: ThemedReactContext) : FrameLayout(context) {
       backgroundColor?.let {
         shape.fillColor = ColorStateList.valueOf(Color.parseColor(it))
       }
+    }
+  }
+
+  private fun setCardBrandTint(color: Int) {
+    try {
+      cardInputWidgetBinding.cardBrandView::class.java.getDeclaredField("tintColorInt").let { internalTintColor ->
+        internalTintColor.isAccessible = true
+        internalTintColor.set(cardInputWidgetBinding.cardBrandView, color)
+      }
+    } catch (e: Exception) {
+      Log.e(
+        "StripeReactNative",
+        "Unable to set card brand tint color: " + e.message)
     }
   }